[SPARK-41768][CORE] Refactor the definition of enum to follow with the code style
