--- core/src/main/scala/org/apache/spark/status/protobuf/DeterministicLevelSerializer.scala
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.status.protobuf
-
-import org.apache.spark.rdd.DeterministicLevel
-import org.apache.spark.status.protobuf.StoreTypes.{DeterministicLevel => GDeterministicLevel}
-
-private[protobuf] object DeterministicLevelSerializer {
-
-  def serialize(input: DeterministicLevel.Value): GDeterministicLevel = {
-    input match {
-      case DeterministicLevel.DETERMINATE =>
-        GDeterministicLevel.DETERMINISTIC_LEVEL_DETERMINATE
-      case DeterministicLevel.UNORDERED =>
-        GDeterministicLevel.DETERMINISTIC_LEVEL_UNORDERED
-      case DeterministicLevel.INDETERMINATE =>
-        GDeterministicLevel.DETERMINISTIC_LEVEL_INDETERMINATE
-    }
-  }
-
-  def deserialize(binary: GDeterministicLevel): DeterministicLevel.Value = {
-    binary match {
-      case GDeterministicLevel.DETERMINISTIC_LEVEL_DETERMINATE =>
-        DeterministicLevel.DETERMINATE
-      case GDeterministicLevel.DETERMINISTIC_LEVEL_UNORDERED =>
-        DeterministicLevel.UNORDERED
-      case GDeterministicLevel.DETERMINISTIC_LEVEL_INDETERMINATE =>
-        DeterministicLevel.INDETERMINATE
-      case _ => null
-    }
-  }
-}
